refactor: ファイルの設計を変更
TodoItems.tsxに対してきちんと単一責任の原則を反映させるため
