+++ todo-app/src/App.tsx
@@ -1,5 +1,15 @@
-import InputBar from "./component/InputBar";
+import React, { useState } from "react";
+import InputBar from "./components/InputBar";
+import TodoItems from "./components/TodoItems";
 
-const App = (): JSX.Element => <InputBar />;
+const App = (): JSX.Element => {
+  const [todoItems, setTodoItems] = useState<string[]>([]);
 
+  return (
+    <>
+      <InputBar todoItems={todoItems} setTodoItems={setTodoItems} />;
+      <TodoItems todoItems={todoItems} />
+    </>
+  );
+};
 export default App;
